The E8 lane joins signal to send, on three clocks, behind a cost fence
THE MISSING MIDDLE IS BUILT. The adapter could place an order and the guard could
refuse one, and nothing joined them. `prop/e8_executor.py` does, and it is the
only file in the lane that decides to trade. Verified end to end against the live
account: 20 sleeves considered, $50 risk each, guard passed at $100,000 equity
with $2,500 to today's floor, 0 signals -- correct, the market is shut and the
last closed bar is Friday's.

THE BOOK IS 20 SLEEVES, NOT 24, and the venue decided that. `prop/e8_book.py`
selects venue-listed first, then round-robins across mechanisms, then ranks by
expectancy within one, and takes one row per (mechanism, symbol) because two
certificates on the same mechanism and instrument are one bet with two names.
Against E8's live catalogue that yields exactly 20: discovered 12,
session_range_breakout 5, overnight_gap_decay 3, mean ev 0.2714, 20 distinct
symbols. Ten certificates are blocked by the venue and 31 are duplicates.

SIZE IS 0.05%, RE-DERIVED UNDER TAIL CORRELATION, and this is the number that
changed. The earlier 24 x 0.07% came from a LINEAR rho. `orthogonality.py`, once
its provenance regression was fixed, measures mean pairwise pearson 0.155 against
a tail-implied 0.251 -- a 1.62x multiplier -- and the 12/5/3 mechanism mix puts
this book's structural rho at 0.408, so about 0.66 in the decile that decides
whether the daily floor is touched:

  20 sleeves   risk    linear 0.408      TAIL 0.66
               0.07%   94.9% (63d)       87.4% (53d)
               0.05%   93.9% (89d)       90.2% (78d)

0.05% is the size that survives being wrong about the correlation, which is the
only assumption here never measured on live fills. Honest headline: ~90% pass,
~78 days median, 1.0% gross simultaneous exposure.

THE COST FENCE IS WHAT MAKES ARMING SAFE WHILE THE SPREAD IS STILL UNKNOWN. The
account is commission-free, which means the cost is a wider quote rather than
none, and it has never been measured during the hours this book trades -- the
only reading so far was 23:56 on a Saturday with the quotes frozen. The desk has
independently measured on Fusion that AUDNZD widens 88x at hour 00 against a
mechanism that dies at 2.34x cost, and every certificate here fires in the asia
window, which opens directly after the rollover. So the executor REFUSES any
sleeve whose round-trip spread exceeds 25% of its own stop distance. A
certificate's edge is denominated in R; paying a quarter of an R to enter is not
a smaller edge, it is a different trade from the one that was certified. It is a
refusal and never a resize -- sizing down to absorb a bad spread keeps the trade
and hides the cost -- and each refusal is named with the number that caused it,
so a night of them is a spread measurement rather than a silence. Per-sleeve and
per-pass, so a symbol fine at 02:00 and impossible at 00:00 trades at 02:00: the
"different hours, not a different size" answer applied automatically.

THREE CLOCKS, registered by install_e8_tasks.ps1:

  E8-Executor  hourly at :02   two past, because the families read the LAST
                               CLOSED bar and a pass at :00 races the close
  E8-Book      daily at 23:40  an instrument E8 delists stops being traded the
                               next day rather than becoming an hourly rejection
  E8-Spreads   at boot         the cost measurement every pass number depends on

ARMING IS A FILE, NOT A FLAG. data\E8_ARMED present means send; absent means do
everything else and write the same ledger. The kill switch has to be something a
person can operate in one action, from a file browser, at three in the morning,
without editing a scheduled task -- GENERIC_EXEC_ENABLED works the same way for
the same reason. Disarming closes nothing: every order this lane sends carries
its stop AT THE VENUE, and the guard still flattens on a stand-down or a breach
because those run regardless of arming. Currently SHADOW.

Sizing refuses to guess an FX rate: a non-USD quote currency with no measured
rate takes the venue minimum and says so in the basis, which is under-sized and
never over. Contract size comes from the venue or the order is not derived at all.

Gates: lint, compile, collect, types all green.

+++ desks/mt5/prop/e8_guard.py
@@ -167,7 +167,7 @@ def assess(equity: float, *, now: datetime | None = None, state: dict[str, Any] 
     # floors, where the error runs the other way and would report OK one hundredth of a cent
     # inside a breach. Money is counted in cents; the comparisons are made there.
     def _c(x: float) -> int:
-        return int(round(x * 100))
+        return round(x * 100)
 
     day_start = float(st["day_start"])
     day_pnl = equity - day_start
